added: 1) fragment showing the weather by day 2) refresh layout, for update weather fragment

+++ app/src/main/java/android/yushenko/openweather/model/Current.java
@@ -33,7 +33,7 @@ public class Current {
     private Double dewPoint;
     @SerializedName("uvi")
     @Expose
-    private Integer uvi;
+    private Double uvi;
     @SerializedName("clouds")
     @Expose
     private Integer clouds;
@@ -117,11 +117,11 @@ public class Current {
         this.dewPoint = dewPoint;
     }
 
-    public Integer getUvi() {
+    public Double getUvi() {
         return uvi;
     }
 
-    public void setUvi(Integer uvi) {
+    public void setUvi(Double uvi) {
         this.uvi = uvi;
     }
